working on tradingService test

+++ java/swe1/MyServer/src/main/java/at/fhtw/bif3/domain/TradingDeal.java
@@ -20,6 +20,6 @@ public class TradingDeal {
     @SerializedName("CardType")
     private CardType cardtype;
     @SerializedName("MinimumDamage")
-    private int minimumDamage;
+    private double minimumDamage;
     private User creator;
 }
